server/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2
django-cors-headers==4.3.1
django-debug-toolbar==4.1.0
django-filter==24.2
django-phonenumber-field==7.3.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
drfpasswordless==1.5.9
frozenlist==1.4.1
gunicorn==22.0.0
httpagentparser==1.9.5
idna==3.7
inflection==0.5.1
install==1.3.5
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
multidict==6.0.5
packaging==24.1
phonenumbers==8.13.37
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python-http-client==3.3.7
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sendgrid==6.11.0
shortuuid==1.0.13
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
twilio==9.1.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
